Wait for AAudio state changes
From certain crash reports, it appears that we are getting calls to
OnAudioDataRequestedCallback() after destroying AAudioOutputStream. This
shouldn't happen, but there might be underlying issues in AAudio.

This CL waits for our AAudio stream to transition to a target state,
before allowing the stream to be destroyed. This means waiting for
STOPPED and CLOSED when stopping and closing.

Change-Id: I0e7fb332a1da9b4709a14b3206503be8434af534
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2340061
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Thomas Guilbert <tguilbert@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#796127}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9503b61e108c74a12252fab504dc1ff4b97fd7a0