media/gpu/video_encode_accelerator_tests: Don't read many frames in a test file
The test file input to video_encode_accelerator_tests can have a
lot of video frames. Since the tests keeps all the frames in
shared memory or with file descriptor, the test suffers from the
many video frames. In fact, with tulip2 test file
AlignedDataHelper fails on creating shared memory because the
opened shared memory is more than threshold on some systems.
This CL limits the maximum frames to be used in a test file to
avoid this problem.

Bug: 1045825
Test: video_encode_accelerator_tests in tulip2-320x180.vp9.webm
Change-Id: I9862180be662ba463fc39a47a6c12d7f63d55718
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2336318
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: David Staessens <dstaessens@chromium.org>
Cr-Commit-Position: refs/heads/master@{#798028}
GitOrigin-RevId: df6bceccfd63cf83d28212b89eef3461fa195198