Merge invisible VP9 frames into super frames for macOS VP9.
This assembles superframes from vp9 alt ref frames since that's
what the macOS VP9 decoder is expecting. Where possible it tries
to avoid copying the input buffer, but when we need to create a
super frame, a super block is allocated for the combined frame.

Fixed: 1115334
Change-Id: I9bfe08201fea1b91f6416143d180bf3d78164827
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2354641
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#798408}
GitOrigin-RevId: 71d8568305da075bf8282f3baf72212aca923b19