media/VaapiVideoDecoder: refactor failures UMA
ToT VaapiWrapper has a simple report-to-uma callback that clients have
to implement and that usually just points to "error happened" in either
the normal or the Vpp VaapiWrapper.

This CL changes that:
- VaapiWrapper provides a function for this: ReportVaapiErrorToUMA().
- Clients bind their histogram (eg. Media.VaapiVideoDecoder.VAAPIError)
to that new ReportVaapiErrorToUMA() method.
- A new VaapiFunctions enumeration (with a matching UMA enum) is added
to vaapi_wrapper.cc to identify which concrete vaBla() function failed,
and is logged to UMA.

This forces ToT UMA histograms (and associated enums) to be marked as
obsolete, namely:
-Media.VAJDA.DecoderFailure
-Media.VAJDA.VppFailure
-Media.VAJDAWorker.DecoderFailure
-Media.VAJEA.VppFailure
-Media.VAVDA.DecoderFailure
-Media.VAVEA.EncoderFailure

New ones are landed instead (with associated VaapiFunctions enum):
-Media.VaapiImageDecodeAcceleratorWorker.VAAPIError
-Media.VaapiJpegEncodeAccelerator.VAAPIError
-Media.VaapiJpegEncodeAccelerator.VAAPIVppError
-Media.VaapiMjpegDecodeAccelerator.VAAPIError
-Media.VaapiMjpegDecodeAccelerator.VAAPIVppError
-Media.VaapiVideoDecodeAccelerator.VAAPIError
-Media.VaapiVideoDecoder.VAAPIError
-Media.VaapiVideoEncodeAccelerator.VAAPIError
-Media.VaapiVideoEncodeAccelerator.VAAPIVppError

This CL also adds two new boolean histograms for VaapiVideoDecoder,
who had none.

Bug: b:162962069
Change-Id: I561e48183f61de86247598fea4f3b10cd590c2e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2339787
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799389}
GitOrigin-RevId: 993d2e9e01bdbe3968107250137a686a138ab082