media/gpu/vaapi/vaapiWrapper: Fix kVaapiFunctionNames initialization
The commas are forgot in the kVaapiFunctionNames initialization.
Thus it initializes the first element and the second or later
element is nullptr. SIGSEGV is caused in attempting to log the
name because the char's pointer is nullptr.

Bug: b:162962069
Test: video_encode_accelerator_tests on atlas
Change-Id: Ifbb081b8e92350df5b83510087a042b7213bc377
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2362378
Auto-Submit: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799477}
GitOrigin-RevId: 254b9adf837e6cc93b4b8b91a0e3c4f98aecfe18