media/gpu/{SSIM/PSNR}VideoFrameValidator: Add an option to evaluate with the average
Encoded bitstreams can contain frames that are worse than the
specified tolerance, while the quality of the overall stream is
good. To evaluate with the overall quality, this CL adds an option
to PSNR and SSIMVideoFrameValidator that the validation fails
when the average of the values are worse than their tolerances.

Bug: 1045825
Test: video_encode_accelerator_tests on rammus
Change-Id: I3d68616081d00a164d0495ed9b295bc6afa0e235
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2359876
Reviewed-by: David Staessens <dstaessens@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799946}
GitOrigin-RevId: a691e243c114459a2545fbf5b0eef8fb911505ea