[Fuchsia] Provide cdm_data_directory to FuchsiaCdmManager
The FuchsiaCdmManager needs a FilePath to write CDM user data to. The
Fuchsia ContextProvider receives the persistent directory via the
CreateContextParams::cdm_data_directory field. This CL takes the
cdm_data_directory and mounts it to /cdm_data. This path is provided to
the FuchsiaCdmManager, which does not use it yet.

Bug: 991723, b/162078899
Change-Id: I8c6ba6236deb936f743fb70a7f22779e0d3802c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2330811
Commit-Queue: RJ Ascani <rjascani@google.com>
Reviewed-by: Wez <wez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#801467}
GitOrigin-RevId: 2dfa8cd820c255932a9579eaeaee01e1420e4700