media/gpu/test/VFValidator: Workaround the buffer synchronization problem on intel 3.18
There is a synchronization problem on the mapped buffer on intel
skylake devices. This is due to some missing 9165 patches in kernel
3.18, and thus will be fixed if the kernel is upreved to 4.4 or
newer. This bug causes a decoded content validation failure in
video_decode_accelerator_tests.
This is a workaround to this bug; usleep() is added after the
buffer is mapped and before the frame validation if the kernel
version is 3.18 and the processor is Skylake.

Bug: b:149808895
Test: video_decode_accelerator_tests on sentry
Change-Id: Ib77ebe0b83834e4d91bc994cbe6912348862e021
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2371724
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#801614}
GitOrigin-RevId: 276dd2ba667d688e3c679c63857742c48e21fe9f