Have WebMediaPlayerImpl inform delegate when output device changing is unavailable
This change has the WebMediaPlayer register a callback with the
WebAudioSourceProviderImpl to be called when it is no longer possible to
change audio output devices.

When the audio output device cannot be changed, the Global Media
Controls UI should not display a device picker. This change is a part
of a series of changes for appropriately hiding this UI.

Bug: 1120620
Change-Id: I9b65a78e57bc08358b5cfc0ff46733ca545df825
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2369439
Commit-Queue: Noah Rose Ledesma <noahrose@google.com>
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#802302}
GitOrigin-RevId: 3cf747620338d685cce34edd923c0eff77b8a8d5