Add blink::Picture::ScopedSharedImage
Create a blink::Picture::ScopedSharedImage structure, which will
keep a SharedImage alive as long as it is alive. This will be
used to pass a SharedImage from the hardware decoder across to
the VTVideoDecodeAccelerator.

The PictureBufferManager is where the decision about use of
SharedImage will be made. It will be sent to the
VTVideoDecodeAccelerator by specifying an empty PictureBuffer.

Bug: 1108909
Change-Id: Iabb71b1b69dd922dc44aca3821f243fcdd22c6d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2377419
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#802723}
GitOrigin-RevId: 5e3de235f357e9c1bec7e9e03b37ad13c470a10c