Add SharedImage support to VTVideoDecodeAccelerator
Directly create a SharedImageBackingGLImage that wraps the
GLImage is already being created. Register this backing
using SharedImageFactory::RegisterBacking. This is the same
pattern used by GpuSharedImageVideoFactory::CreateImageInternal
among others.

Pass the reference-counted destructor callback to PictureReady
callback.

Move the format definitions to a single block and add a TODO
indicating that the format of GL_BGRA seems inappropriate (but
it works, for unclear reasons).

Assume that the SkiaRenderer is only used with the passthrough
decoder. Either the decoder type will need to be plumbed, or
we will bake in the assumption.

Add GPU_GLES2_EXPORT to the relevant structures used in media
code.

Bug: 1108909
Change-Id: I33e230945e9c3d231195f735a878d903f11c95ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2380743
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#803515}
GitOrigin-RevId: 78fc7422e5232c3cd7bc99abefe871b8597da481