media/gpu/vaapi: Introduce ScopedVABuffer and use for encoding and Vpp
VaapiWrapper owns VABufferIDs of the VA buffers that are used
outside of VaapiWrapper. This CL moves the ownership out of
VaapiWrapper. The VA Buffers are owned by ScopedVABuffer in order
to prevent the buffer leakage.
Ths ScopedVABuffer will be used for other VABufferIDs later.

Bug: b:166646505, b:162962069
Test: video_encode_accelerator_tests on atlas
Test: media_unittests --gtest_filter=VaapiVideoEncodeAcceleratorTest
Test: vaapi_unittest --gtest_filter=VaapiUtilsTest
Change-Id: I7b0dda7491d29ebba70bc42051a8657144209dda
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2379423
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#804103}
GitOrigin-RevId: e97a414a890736eeef6ee29535f36feabda5b23d