Boost thread priority while loading COM VideoCapture
The thread priority boost need to be apply at each call.

This is a follow up on:
  https://chromium-review.googlesource.com/c/chromium/src/+/2378395

Noticed in b93726646e382c66
The crash shows that the SCOPED_MAY_LOAD_LIBRARY_AT_BACKGROUND_PRIORITY()
already got called. The thread is still at a background priority and
it is loading a DLL.

R=joenotcharles@chromium.org,guidou@chromium.org

Bug: 973868
Change-Id: Ib4f0eaf317b8b6408f1f4ede8607ff00c1b5cc5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2402163
Reviewed-by: Joe Mason <joenotcharles@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#805527}
GitOrigin-RevId: 116d364827978ac5c987a90b329f3ca3f01a8a04