media/gpu/vaapi: batch submitting VABufferIDs for MJPEG decoding
This CL follows crrev.com/c/2393629 by batching up SubmitBuffer
calls in the JPEG decode accelerator, used e.g. for MJPEG camera
stream decoding.

This was verified via chrome:tracing while capturing HD on kohaku
and reks (Braswell) in the link [1]. Similarly to what is seen in
the previous CL, the combination of SubmitBuffer() calls on ToT
is ~10% slower than the batched SubmitBuffers(). A key difference
is that since JPEG is way simpler than e.g. VP9, the submitting
time is comparable to the ExecuteBuffers_Locked(), so this CL
has some minor overall decode-time savings. But still the target
here is to reduce contention on the lock and libva for video
conferencing cases with multiple remote participants a la Meet.

[1] webrtc.github.io/samples/src/content/getusermedia/resolution/

Cq-Depend: chromium:2393629
Bug: b/166646505
Change-Id: I183cc151de24d8fdbdc62e0172be296dacbf26f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2393525
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Reviewed-by: J Kardatzke <jkardatzke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#805557}
GitOrigin-RevId: 405ed4acb36bf83beed6a5ada825179389a3c586