Suggest --no-sandbox if fake audio capture fails reading the file
When simulating a fake audio device with the
--use-fake-device-for-media-stream
--use-file-for-fake-audio-capture=audio.wav
arguments, the filepath (`audio.wav` here) isn't accessible because
of the mac sandbox. As a result, opening the file fails and the audio
file isn't played.

This changeset adds a suggestion to disable the sandbox if
simulating a fake audio device fails reading the audio file.

Bug: 1032604
Change-Id: Ia4bc9a1ba4eb9a6d199f7d018bb0122bbd912517
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2383230
Reviewed-by: Tommi <tommi@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#805945}
GitOrigin-RevId: 167f34ae932ed8f702d9127339acbff246b4eb47