VideoCaptureDeviceAVFoundation: move static functionality to a new utils file.
This change prepares for development of capturer features behind a new flag. Moving static functionality into a separate file allows the capturer implementation to be duped, and the creator to create either implementation based on a flag.

Future CLs will add the flag and perform the duplication.

Bug: chromium:1126690
Change-Id: I8592a2ba81765e999c497eda9a6f09cc644e1423
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2401520
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: ccameron <ccameron@chromium.org>
Commit-Queue: Markus Handell <handellm@google.com>
Cr-Commit-Position: refs/heads/master@{#805965}
GitOrigin-RevId: f4b4b2d0c48c128da7e5345b809b381601bc854a