media: Add KeySystemConfigSelector::Status
This CL refactors how KeySystemConfigSelector::SelectConfig() returns
the results. A Status is added to provided more details on the config
is not supported. The promise rejection messages are also updated to
return different strings on different statuses, which could be useful
or troubleshooting.

Currently we still return the same message for kUnsupportedKeySystem and
kUnsupportedConfigs to avoid updating many tests and potentially affect
existing JS application. In the future we may consider returning
different strings for that case.

For kUnsupportedPlatform we return a different string. This should be
fine for privacy because it'll only be returned when the platform does
not support the key system at all and should not reveal any user/device
information.

Bug: 760720,1126664
Change-Id: I8df167e5c3e86607917d5414a558aae32eed027c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2405734
Reviewed-by: John Rummell <jrummell@chromium.org>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#806255}
GitOrigin-RevId: 63da018ee845021c9da97852371add2902bb7225