media/gpu/v4l2vd: let the pipeline also adjust the decoder buffer size
When setting a new resolution, the V4L2 VD calls the pipeline's
PickDecoderOutputFormat(), which is in charge of picking an output
format that we can eventually display, if needed with the help of an
image processor. However the image processor's requirements in terms of
buffer size might be different from those of the decoder.

Make PickDecoderOutputFormat() return the adjusted size along with the
selected pixel format, so the decoder can align to it and set a
resolution that the image processor can use as input.

BUG=b:168193954
TEST=video.DecodeAccelVD.h264_resolution_switch passes on hana-kernelnext.

Change-Id: Id92e75b3d8a3b0e1291f0638fb41ddd5956e6879
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2405347
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Auto-Submit: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#806527}
GitOrigin-RevId: 0276e8f329b10e6ef4f60a9975925ef62feec1ee