media/gpu/vaapi: batch submitting VABufferIDs for JPEG encoding
This CL follows crrev.com/c/2393629 by batching up SubmitBuffer
calls in the JPEG encode accelerator (JEA).

This was not verified perf-wise but the savings are obvious given
all the other CLs on this bug: we avoid locking-unlocking the
|va_lock_| for every buffer submission, and there's 5 of them
per JEA Encode() call. Time saving of course will be minimal but
the target here is to reduce contention on the lock

Bug: b/166646505
Change-Id: I4ad8ba3498cdcb1d1259dc527a7b4799320fc599
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2405417
Auto-Submit: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#807039}
GitOrigin-RevId: e47c7b77b1e4563b02f849d42487d07d553bb6cb