media/gpu/v4l2: add V4L2Queue::TryFormat()
We were using SetFormat() when we wanted to just test the availability
of a given format on a device. This is a bit misleading for the reader.

Implement and use TryFormat(), which properly uses the TRY_FORMAT ioctl,
and use it where relevant.

BUG=None
TEST=video.DecodeAccel.h264 passed on Trogdor

Change-Id: I223a2911451bae753e8560180ccdaa5a3502a1db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2422103
Commit-Queue: Fritz Koenig <frkoenig@chromium.org>
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>
Auto-Submit: Alexandre Courbot <acourbot@chromium.org>
Cr-Commit-Position: refs/heads/master@{#809404}
GitOrigin-RevId: ae212018b4b3a11acd49055745257a5feed8578c