Added UseAfterFreeChecker to see if MediaWebContentsObserver is
We're getting some crashes that look like either use-after-free or
heap corruption in MediaWebContentsObserver.  This CL adds a quick
check for it, that will abort if it finds either, and tell us which
one it found.  Hopefully, it will help us to narrow down if (a)
use-after-free/corruption is actually what's going on, and (b) if
it's happening as part of the call into MWCO or earlier.

Change-Id: I91edc46af55b9dcbd223b01c3bb80c22b1386903
Bug: 1111291
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2429246
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Frank Liberato <liberato@chromium.org>
Cr-Commit-Position: refs/heads/master@{#810365}
GitOrigin-RevId: 0b1d4ce6f2b09e79e3251f6920c85ca3d452c2c5