Remove SurfaceEmbeddingTime and LocalSurfaceIdAllocation
The Histogram for SurfaceEmbeddingTime is being marked obsolete. It has
been supplanted by Graphics.Smoothness.* metrics for on going monitoring.

The trace events which were added to supplement the metric are still
very useful for debugging and will remain.

LocalSurfaceIdAllocation was used to pipe the allocation time along with
the LocalSurfaceId that was needed for the actual Surface Sync. This
change also removes that, leading to a mechanical refactor of classes to
use just LocalSurfaceId directly.

TBR=boliu@chromium.org
TBR=dalecurtis@chromium.org
TBR=dcheng@chromium.org
TBR=dgozman@chromium.org
TBR=guidou@chromium.org
TBR=mbarbella@chromium.org
TBR=oshima@chromium.org
TBR=seantopping@chromium.org
TBR=sky@chromium.org
TBR=yfriedman@chromium.org

Bug: 1000081
Change-Id: Ibe9a85d63b534c55548a109ea746764d8c66f2b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2415128
Reviewed-by: Sean Topping <seantopping@chromium.org>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Martin Barbella <mbarbella@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Reviewed-by: kylechar <kylechar@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/master@{#810418}
GitOrigin-RevId: d0345bfbf2567baa499d4428eb6b875c31a6a817