Enable SharedImages for VTVideoDecodeAccelerator
Add a method to media::VideoDecodeAccelerator to indicate whether
SharedImage-backed picture buffers are supported. Make this return
true only for the VTVideoDecodeAccelerator.

Add a parameter to media::PictureBufferManager::CreatePictureBuffers
to indicate if SharedImage-backed PictureBuffers are desired.

Hook the two of these together in the ProvidePictureBuffersAsync
method of media::VdaVideoDecoder.

Video playback fails with the Metal-based command decoder, because
it binds IOSurfaces as TEXTURE_2D, not TEXTURE_RECTANGLE. Plumb
through a flag to specify if TEXTURE_RECTANGLE support is present.

Bug: 1108909
Change-Id: I1e42aa38e122714517724cee87eb51987645e410
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2427545
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: ccameron <ccameron@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#811077}
GitOrigin-RevId: 8b7817c71c6557ed5895916a2f61517410f9afa3