media/gpu/vaapi/VP9Encoder: Enable error_resilient_mode in temporal layer encoding
A frame context is necessary to parse vp9 compressed data and
the context is updated every frame by default. In other words,
it is necessary to see all previous frames.
In temporal layer encoding, frames in temporal layers, e.g. TL0
only stream and TL0+TL1 stream, must be able to be decoded even
though there is no frames in TL2. To resolve the frame context
problem, we enable error resilient mode in temporal layer
encoding. It let the frame context not be updated and thus the
syntax of a frame can be decoded independently of the previous
frames.

Bug: b:169219904, b:169538499
Test: https://appr.tc/?vsc=vp9&vrc=vp9
Change-Id: I976dc2d2f0f91d69b87b081da22f1c153ecb634c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2435030
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#811421}
GitOrigin-RevId: 337d68d7ac3792d369b27a9f46e8ecefb1df71c9