[macOS Capture] Add PixelBufferPool and tests.
The pixel buffer pool is a light-weight wrapper for CVPixelBufferPool.
It takes care of configuring the pool, which avoids working with some
cumbersome Core Foundation objects.

This CL also adds tests to make sure the pool is behaving as expected,
since the Apple documentation is rather poor in details.

The intent is to use the PixelBufferPool in future CLs in order to
use the VTPixelTransferSession for more efficient (hopefully) scaling
and pixel format conversions.

// Guido is OOO today so TBR-landing for BUILD.gn changes.
TBR=guidou@chromium.org

Bug: chromium:1132299
Change-Id: I28ffa7d75dbffd4fcf44c7fb2b22e89e19d6a0ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2431427
Reviewed-by: Henrik Boström <hbos@chromium.org>
Reviewed-by: Markus Handell <handellm@google.com>
Reviewed-by: ccameron <ccameron@chromium.org>
Commit-Queue: Henrik Boström <hbos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#811650}
GitOrigin-RevId: 5320dd38c134ed6bd1c50202e61d92987311c890