Add media::GetFormatDescriptionColorSpace
This function is nearly identical to media::GetImageBufferColorSpace,
but instead operates on a CMFormatDescriptionRef.

Add DCHECKs that the behavior of the two functions be the same, that is,
the kCVImageBuffer-prefixed and  kCMFormatDescription-prefixed
constants be equal.

Use base::mac::CFCast instead of reinterpret_casts.

The next patch in this sequence will move these two functions over
to media/base/mac, because it will be accessed in capture.

Bug: 959962
Change-Id: I8a4d8468abc866a13cb6181cbd49d9ee04288124
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2436513
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#811829}
GitOrigin-RevId: 6c5fa707365646a3c2b8cae4e20f5a5f333d5dac