media/gpu/v4l2vea: use buffer affinity tracker
When using DMABUFs, it is preferable to use the same V4L2 buffer with
the same underlying buffer, as failure to do so results in memory
unmappings/remappings in the driver. Use the newly introduced buffer
affinity tracker and V4L2Queue::GetFreeBufferForFrame() method to
achieve this transparently in the video encoder.

BUG=b:159688625
BUG=b:167412992
TEST=video.EncodeAccel.h264_360p_i420 passes on Kukui.

Change-Id: I050c29ecf8b364869f687c3d8632718aef868144
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2415910
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#813514}
GitOrigin-RevId: b2a7b0cc4ec6752e1ada55112ceefa2097f3d364