Don't invalidate WeakPtrs for kConfigChange; remove state.
The way the kConfigChange state was used is not safe and ultimately
useless. It was set atomically on the main thread, but operations
on the decoder thread would query multiple times within a single
function call and were not expecting kConfigChange. It's ultimately
useless since we just end up stopping the decoder thread regardless.

Except for cases of destruction or error, it's incorrect to invalidate
the WeakPtr posted tasks too. These manage picture buffer allocation
and client notifications that should never be dropped.

This also applies a temporary fix for a DCHECK() crash and removes
some stale debugging code.

Bug: 1111273, 1139489
Change-Id: I42a9c76f7ea272c57b8c7aa3e78bd379c3120c61
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2481945
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Cr-Commit-Position: refs/heads/master@{#818554}
GitOrigin-RevId: 0dbad76d85cedc53ed428a93c71a60b8c79ca481