[Fuchsia] Allow SetVolume on uninitialized FuchsiaAudioRenderer
initialized. FuchsiaAudioRenderer was not handling this correctly.
It was trying to call AudioConsumer.BindVolumeControl before the
AudioConsumer is connected. As result it was logging an error
message about disconnected VolumeControl without setting the volume.
Updated SetVolume() implementation to persist the volume value and
send it as soon as AudioConsumer is connected.

AudioRenderer: :SetVolume() may be called before the renderer is
Change-Id: Ic5418c2c002c1181da01e8a278b0b53845baee47
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2481958
Auto-Submit: Sergey Ulanov <sergeyu@chromium.org>
Commit-Queue: Kevin Marshall <kmarshall@chromium.org>
Reviewed-by: Kevin Marshall <kmarshall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#818564}
GitOrigin-RevId: 4767de3a482f0a7e9b3cdbab076875ecff0ef7d7