Fixing the RawInputDevice removal logic in UserInputMonitorWin
This fix is a result of an investigation to find out why KeyboardLock
stops working after approving the microphone permission on another
website (even after it has been navigated away from or the tab closed).

I tracked the issue down to this class as it registers a RawInputDevice
to listen for keyboard input but failed to unregister it properly when
the monitor was stopped.

Per MSDN, the HWND param should be NULL when unregistering the device:
https://docs.microsoft.com/en-us/windows/win32/api/winuser/nf-winuser-registerrawinputdevices

Change-Id: Icf54f3ac61fe5020b00472e6680e4e533b1fd256
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2482044
Reviewed-by: Chrome Cunningham <chcunningham@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Joe Downing <joedow@google.com>
Cr-Commit-Position: refs/heads/master@{#818584}
GitOrigin-RevId: e57924dbb155ebd44e7c2ba68f70b935b5a7ff08