Use IOSurface directly in gfx::BufferHandle
Change gfx::BufferHandle to use a gfx::ScopedIOSurface instead of a
gfx::ScopedRefCountedIOSurfaceMachPort.

Reason for using a mach port: It's the thing that goes over all forms
of IPC.

Reason for not using a mach port: The mach port's existence makes the
IOSurface report that it is in use. This makes it so that things like
CVPixelBufferPools will not re-use that IOSurface. The capture pipeline
keeps lots of gfx::BufferHandles around, which causes the underlying
CVPixelBufferPool to go haywire (in this case growing without bound,
but another valid way to blow up is to just hang and never capture
any more frames).

Also remove some debugging code.

Bug: 1139105
Change-Id: I1ffa11a325b6a8e1811a23b1c4d7dac00fad49db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2485992
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Peng Huang <penghuang@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#818727}
GitOrigin-RevId: 48792632e7aa196b15a0252ea0b07d2e574522a9