[ozone/x11] Enabled UserInputMonitor.
The UserInputMonitorLinux depends on X11 directly and cannot be used by
Ozone platforms because of that.

This CL introduces a new platform interface for the user input monitor,
and moves the X11 implementation to //ui/base/x.  The
UserInputMonitorLinux now uses either the X11 implementation directly
(when X11 is enabled but Ozone is not) or requests the platform
implementation when Ozone is enabled.

Bug: 1116414
Change-Id: I0a21074f2fe028fea13c7efb33ddf9d102c9f35c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2489863
Commit-Queue: Alexander Dunaev <adunaev@igalia.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#821657}
GitOrigin-RevId: 1000bbb0e4900f64c670f7a0281f1611f6145932