Always pass on volume changes to the renderer.
It may be using this information to optimize startup and it may be
lost if not passed on to the renderer.

Specifically in this case the volume change is coming when the
pipeline is in the kResuming state, which it remains in until
CompleteSeek() is called. So volume changes that occur within the
window between InitializeRenderer() and CompleteSeek() are lost.

We could call SetVolume() during CompleteSeek() like we do for
InitializeRenderer(), but just removing the restriction seems
fine since renderers were already expecting the volume state to
change at any time.

R=xhwang

Bug: 1143395, 1140807
Change-Id: Icbe20f35e9abf2b76c189d48efaa85dc5d3b3fac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2508174
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Auto-Submit: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#822355}
GitOrigin-RevId: 7502f1e8bb5241f99101d460c73028e1112d4a13