Move WatchTimeReporter to Blink.
Make the WatchTimeReporter accessible from the Blink component
WebMediaPlayerMs, used in follow-up https://crrev.com/c/2513645 to
report the duration of media streams.

The WatchTimeReporter is placed in Blink common/ in order to be
accessible from both //media/blink/ and Blink renderer/modules, and to
be able to access the MediaMetricsProvider mojom, which cannot be
accessed from Blink public/web.

Bug: 1077707
Change-Id: Id1f52549b6f1e791ffd0e71f7a95f12eb1cc57f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2401459
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Marina Ciocea <marinaciocea@chromium.org>
Cr-Commit-Position: refs/heads/master@{#823483}
GitOrigin-RevId: bcf64a0e094f38573682d95faefcff93b5cde88c