media/gpu/test/video_encode_accelerator_tests: Add NV12 dmabuf scaling test case
This adds a test case of encoding NV12 Dmabuf VideoFrame with
down-scaling; the input resolution is twice larger in width and
height than one configured to VideoEncodeAccelerator. This can
happen in WebRTC simulcast scenario when zero-copy capture
feature is enabled. To not hit the smaller resolution's
limitation, the test case is skipped if the resolution to be
produced is less than 240x180.

Bug: 1045825
Test: video_encode_accelerator_tests on atlas and kukui
Change-Id: I11eac935aa2c6a03bcef0dfa533c8731d593d3df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2342535
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: David Staessens <dstaessens@chromium.org>
Cr-Commit-Position: refs/heads/master@{#827711}
GitOrigin-RevId: 38163d72a9d7fa2383c78055cea488f6ed468dc7