Enable HDR for P010 GpuMemoryBuffers on macOS.
This plumbs HDR support for P010 GpuMemoryBuffers on macOS. It works
by setting up a side channel on the IOSurface for communicating the
MDCV and CLLI SEI data to the final CVPixelBuffer.

This also makes a couple of minor fixes to media/ code:
- VideoDecoderConfig wasn't printing the full HDR metadata.
- VpxVideoDecoder didn't always set the HDR metadata properly.

Note: There are cases where we don't allow GMBs for software decoding:
Full range and odd sized YUV. Possibly we want to range reduce in the
copy to GpuMemoryBuffer process to eliminate the first one.

Bug: 1101041
Change-Id: I3bcc390901020b20514ff18ecc6cc73b59f50525
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2523739
Reviewed-by: ccameron <ccameron@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#827913}
GitOrigin-RevId: a89bd2c6dbf6d0bcb1dccc9057decb54907523c9