media/gpu/v4l2VEA: Fix natural_size check condition
Original V4L2VideoEncodeAccelerator checks that the natural size
of every VideoFrame with the dimension to be encoded. They must
be the same in single cast case, but can be different in
multicast case. The correct check is to compare it with the
natural size of the first VideoFrame. In other words, the
natural_size must be unchanged during encoding in the same
VideoEncodeAccelerator instance.

Bug: b:172223525
Test: meet.google.com on trogdor
Change-Id: I3c21782b86cddaa788d4daf785793f73ec49db9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2541942
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#828641}
GitOrigin-RevId: 0679ab5810be0111dd7c066355a98f750a348e64