Check for output buffers more often in MediaCodecVideoDecoder.
When checking for output buffers while a surface transition is
pending, MCVD will choose to wait if there are still unrendered
output buffers.  However, once the last buffer is rendererd,
MCVD didn't always realize to try again; it did so only if it
ran the codec pump for some reason (e.g., a new DecoderBuffer or
a new codec buffer became available).

This CL causes it to try again whenever an output buffer is
rendered, in case it can make progress because of it.

Bug: 1148995
Change-Id: Ife1ba993424491af4a29c71a979e5bb0c7b79eb9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2547227
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Frank Liberato <liberato@chromium.org>
Cr-Commit-Position: refs/heads/master@{#829315}
GitOrigin-RevId: 3ba8f31f884eeee793f32e726f8f194f37dc8e8e