[macOS Capture] Restore SampleBufferTransformer's pool size to 10.
When the M88 cut was made, the kDefaultBufferPoolSize was 10. While
investigating https://crbug.com/1148964 we disabled and then re-enabled
the InCaptureConvertTnoNv12 feature and along the way changed the pool
size to 5 (trying to see if the issue was related to running out of
handles, but that wasn't the case).

To ensure M88 and Canary uses the same pool size, this CL restores the
old pool value to 10. If we decide to change the pool size we should do
that in a separate easily back-mergeable CL, but there have been no
signs yet that 10 is too many.

Bug: chromium:1148964
Change-Id: Ib122ca1ed9ca061951ee5d77dfa3ddbd84d90f05
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2553548
Commit-Queue: Henrik Boström <hbos@chromium.org>
Commit-Queue: Markus Handell <handellm@google.com>
Auto-Submit: Henrik Boström <hbos@chromium.org>
Reviewed-by: Markus Handell <handellm@google.com>
Cr-Commit-Position: refs/heads/master@{#830143}
GitOrigin-RevId: 0d109814f8ebced275c8f9da8a132dc90cc4384a