Handle PixelBuffers with padded strides in SampleBufferTransformer
The libyuv mode previous only worked with block data, and not with data
that was non-contiguous or had strides larger than the width. This patch
adds tests for padded data, and handles them by using specific
conversion functions instead of the general libyuv::ConvertToI420.

This patch also fixes a bug in the libyuv path where an PixelBuffer not backed by an IOSurface but that didn't need scaling would trigger a DCHECK.

Bug: 1151936
Change-Id: I7bf73871edf2c528c992fb2e0acb1ab7393ff72b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2563718
Reviewed-by: Henrik Boström <hbos@chromium.org>
Commit-Queue: Evan Shrubsole <eshr@google.com>
Cr-Commit-Position: refs/heads/master@{#832758}
GitOrigin-RevId: 0fca5569488e7f0e129781babbbf450b72627760