Pass Pipeline Status Across Mojo Renderer OnError Boundary
This change updates the Mojo Renderer OnError method to take in a
media::Status object, wrapping a PipelineStatus, in order to pass the
error code from the Mojo Renderer Service to the Mojo Renderer.

This was a basic missing bit of functionality that will allow Mojo
Renderers to communicate a more accurate error state for telemetry &
logging purposes.

The media::Status object was used as part of a longer term plan to
allow passing a render or platform specific error status (such as an OS
native status code, e.g. an HRESULT on Windows). A part 2 change will
move the Renderer OnError method to replace PipelineStatus with the
Status object but preserve the existing PipelineStatus enumeration
values & strings for telemetry & logging reporting purposes.

Bug: 1146085
Change-Id: I22faa88cb30e86dceda5a2630d416a62fb0e66ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2553920
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Reviewed-by: Ted Meyer <tmathmeyer@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: William Carr <wicarr@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#833424}
GitOrigin-RevId: ab6a4c33a5153278a33c9b5f6ebdab12304e97a0