media: Fix crash when releasing decoder with ChromeOS CDM
When the VaapiVideoDecoder was released, if it was using the CdmContext
it would have a callback associated with the CDM. Upon destruction of
the decoder, that callback would be destroyed automatically which makes
a call into the CDM.  The CDM is being destructed on another thread, so
there was a race and this could cause a crash.

The fix is to maintain a CdmContextRef in the decoder which keeps the
CDM alive until the decoder is destroyed.

BUG=b:153111783
TEST=Page refresh 10 times in a row w/ CDM did not crash

Change-Id: I0c6f19753b324e7d89db0bd79f6108d5de27ffa3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2572262
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#833863}
GitOrigin-RevId: 883c24e8db59cad2baa5b6a3c8f8143fe4bf195f