vaapi: VP9 accelerator/decoder fixes for protected content
This doesn't fully fix VP9 decrypted decoding, but it makes the
VP9Decoder changes cleaner than they were. It also adds a call that was
missed in the accelerator for setting the crypto params if we seek back
to clear content.

This still has corruption on screen, especially if there's UI activity
over the video. But still not sure how that matters.

BUG=b:153111783,b:155509231
TEST=VP9 decrypt+decode works mostly

Change-Id: Ib2ee90e6ce8fb6f2087c595e1a765e0481cc6249
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2567150
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#833944}
GitOrigin-RevId: 4ed4e33361833ee48ce0b75bc3011664c2f24da1