vaapi: Establish protected session before decode starts
This resolves issues with the driver misinterpreting what frames
are encrypted during the clear lead. The main effect is now we will only
be using protected surfaces with an active protected session.

NOTE: This breaks full sample encryption, we are sorting that out with
Intel, but that's a tiny corner case compared to the benefit this has so
we will enable this now.

BUG=b:153111783,
TEST=Protected playback works for VP9/H264 subsample

Change-Id: Id189f856a911140c5a4a9b9c57d539eeeb24c901
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2583111
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#835962}
GitOrigin-RevId: b069339e0f29f7d430a399fe090b35afbec5ea33