media/gpu/vea_tests: Skip framerate change and nv12dmabuf test cases if kepler is used
A framerate change test case fails if kepler is used. A rate
controller exists in a firmware in kepler and the kepler
driver implementation delegates the framerate to the firmware.
Since kepler is released long time ago, it is hard to ask the
vendor to fix the rate controller. Hence, we skip the test
case when kepler is used in order to get a green dashboard.
Additionally, we don't support dmabuf input frame in kepler. The
test case is also skipped by this CL.

Bug: b:170790027
Test: video_encode_accelerator_tests on atlas and rikku
Change-Id: I4a2a85eaba840656665b54486db65ed58214b64b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2532140
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Auto-Submit: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: David Staessens <dstaessens@chromium.org>
Cr-Commit-Position: refs/heads/master@{#836005}
GitOrigin-RevId: 40237c8f8d8818f632d1b0e4460415fcbce296d1