Support non-HDR high bit depth content in D3D11VideoDecoder.
Previously we assumed all HDR content is 10 bit and all SDR content
is 8-bit; this generally works well, but now that the accelerators
have a GetBitDepth() method, we should use it.

This change utilizes the same mechanism we do for size changes: we
get a kConfigChange event on the first decode which has the real
size and bit depth information. We still assuem HDR is 10 bit and
SDR is 8bit at decoder creation, but if we're wrong we'll now fix
that choice during the first config change rather than silently
emitting P010 as NV12 /grumble.

R=liberato

Fixed: 1115846
Test: Updated unittests.
Change-Id: Ifa56c65c2fd7bb82d404f6a73a4369b347a7d1df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2590395
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Auto-Submit: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Cr-Commit-Position: refs/heads/master@{#836920}
GitOrigin-RevId: 559f9b0c9320ab9d4b21aefc9ac68005220a0ee7