[macOS Capture] Avoid crashing on MJPEG sample buffers.
The SampleBufferTransformer CHECK-crashes if a sample buffer does not
have a pixel buffer AND is not MJPEG. But the assumption that non-pixel
buffer sample buffers most by MJPEG turns out not to hold, with the
following consequences:

In M89, this causes a CHECK-crash.
In M88, this causes a DCHECK-crash, which in official builds means we
don't crash but rather we most likely fail to decode the buffer and
return null frames. This is believed to cause "no video" issues.

This CL adds the workaround not to use the SampleBufferTransformer if
"sampleBufferLacksPixelBufferAndIsNotMjpeg" is true. So in this edge
case we fall back on the same code path that we run in M87 today.

TBR=handellm@google.com

Bug: chromium:1160647, chromium:1160315
Change-Id: I782a36cf38f5ce96a5010966c8b07aedf1b08efd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2599531
Reviewed-by: Henrik Boström <hbos@chromium.org>
Commit-Queue: Henrik Boström <hbos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#838638}
GitOrigin-RevId: 3254471bb1c4b8174a7091d14944f249c658486c