Handle duplicate vp9 pictures in D3D11VideoDecoder.
Allow duplicate pictures by returning the same ref-counted picture
buffer to the VP9Decoder, with an updated timestamp.

This also replaces the boolean "in client use" with a counter, since
the same PictureBuffer can be out at the client as more than one
VideoFrame now.  We can only re-use it for decoding once all of the
client uses are returned to the decoder.

Bug: 1164463
Change-Id: I93fa076ae4fdd955687122df6c4cb12485cd6a7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2623030
Commit-Queue: Frank Liberato <liberato@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#842317}
GitOrigin-RevId: bd2cb9f56dda1041994f5587a522f56313160af5