Stop dangling pointer from being assigned to AudioOutputDevice::callback_
When we use BackupRefPtr for AudioOutputDevice::callback_, it detects a dangling pointer being assigned to it.
This is because, RenderCallback can be destroyed by the time InitializeOnIOThread gets called on the io thread.
This CL adds a check for stopping_hack_ so we won't assign, when Stop() has already been called.

Bug: 1154812, 1080832
Change-Id: Icb02ed9661ebcd2c0ee8c81888bc68c3d1af83a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2575481
Commit-Queue: Keishi Hattori <keishi@chromium.org>
Reviewed-by: Olga Sharonova <olka@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#843465}
GitOrigin-RevId: 0f8133db9115df91422e883f611e674c9f33688d