vaapi: Fix multi-slice encrypted content
This fixes multi-slice content. The proper way of doing it is to
aggregate all of the encryption parameters into one buffer for all the
slices and send that. Previously it was sending an encryption parameter
buffer for each slice.

BUG=b:153111783
TEST=Encrypted H264/HEVC work with single and multi-slice

Change-Id: I57abbc839d1c6878da7c6a13b8b6da9889329179
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2631466
Auto-Submit: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Cr-Commit-Position: refs/heads/master@{#843874}
GitOrigin-RevId: 30f3cacaaf91fff9f466059a22b317de495c4a09