vaapi: Periodically retrieve keys during protected content
For cases where we are video only decrypt, we need to make calls into
OEMCrypto to get the key material even if we already have it in order to
update the key usage times. The update interval should be ~1 min. This
ensures that we are making that call.

BUG=b:178208439
TEST=Key usage time updates are occurring during Netflix playback

Change-Id: I14d19e4a962c79206a8f41fa2ba5281bd0b95c77
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2644909
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>
Commit-Queue: Jao-ke Chin-Lee <jchinlee@chromium.org>
Auto-Submit: Jeffrey Kardatzke <jkardatzke@google.com>
Cr-Commit-Position: refs/heads/master@{#846326}
GitOrigin-RevId: 0ed369d39fae1b20a293f6f157cf9c49f99eb606