Add decode scaling to VAAPI for protected content H264/HEVC/VP9
For protected content, we must render to an overlay. On Intel TGL, there
is a limitation where we can only downscale a certain amount. With 4K
content, that amount does not meet the requirements of downscaling to
the size of the screen in some instances (e.g. on Voxel), which would
prevent us from showing the entire image (it would be clipped).

To workaround this problem, we can scale the image in advance to the
size of the largest connected display. That way we maximize the amount
of downscaling we have left in the compositor since we will never render
larger than the screen size. Doing it this way in the decoder also
prevents another memory copy like would happen if we used the VPP.

This feature in the VAAPI decoder is called SFC scaling. It works by
giving another surface that the decoded output is scaled to, the full
size decoded surface is also needed because that is used for reference
frames.

This uses a Mojo call to functionality we already had in the browser
process to get the screen size. After much searching, we were unable to
find another existing mechanism for accessing screen size from the
decoders in the GPU process.

BUG=b:153111783,b:174174828
TEST=4K HEVC/H264/VP9 protected content renders fully on Voxel

Change-Id: I4af7eb22628afb3f1783ff5a98b69f5d1bccd591
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2641088
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#847076}
GitOrigin-RevId: 2bee3b7915389f9fc9be5246889324bbba9a62b8