Add is_rtc to VideoDecoderConfig.
Allows a VideoDecoderConfig to be marked as a WebRTC config.  Video
decoders can use this to configure themselves properly for RTC.

VideoDecoders can also specify via the new IsOptimizedForRTC()
method that they are not optimized for RTC streams.  This is the
default.  Only MojoVideoDecoder claims to support RTC.  This is to
prevent Chrome's software decoders from decoding RTC streams via
DecoderStream; while they work, they aren't configured the same as
their RTC equivalents, and might not have the same performance
characteristics as a result.

DecoderSelector uses is_rtc() as well as IsOptimizedForRTC() to
disallow any decoder that isn't optimized to decode RTC, when
choosing for RTC configs.

Bug: 1169422
Change-Id: I09151214b1a14a848103caa8d027669237498865
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2644056
Commit-Queue: Frank Liberato <liberato@chromium.org>
Reviewed-by: Will Cassella <cassew@google.com>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#847678}
GitOrigin-RevId: ed3893635e39b8ef995adf9723c66dfe1c11c3f7