vaapi: Skip applying resolution change if resetting
With the addition of an async call in the middle of
ApplyResolutionChange,a situation can occur where ApplyResolutionChange
is called, Reset is called, and then
ApplyResolutionChangeWithScreenSizes is called and the accelerated
decoder is in an invalid state. We should return instead in that case
and the resolution change will get applied once the decoder detects the
resolution again.

BUG=b:178530841
TEST=Verified playback works when Reset happens between resolution
change calls

Change-Id: Id8998b39aa0bfbec5b68b43c0a898e7eaa119f8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2650144
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#847942}
GitOrigin-RevId: 39f30c2ed452cd57a7faff3b023f5a4f32e62407