vaapi: Recover from protected session loss in decoder
There are cases where we lose the hardware context for protected video
decode, and generally we are able to recover from those by recreating
it. Cases where we can't recover are things like HDCP version
requirements.

This change looks for failures in vaBegin/Render/EndPicture which is a
signal that this loss has occurred. We then check if that loss has
happened, and then we make a single attempt to recover. To recover we
need to destroy the protected session first, then send a message back to
the pipeline that we have lost our state (that causes a Seek operation
back to our current position), and then when decoding resumes we will
recreate the entire protected state in the decoder and be able to
continue playing. If we do not succeed past the first frame, we fail and
do not try to rebuild the protected session again for that decode.

BUG=b:178533013
TEST=Playback recovers after PM/hotplug with protected content

Change-Id: Id212cb254627c1ec13a5b84fd97fa81b938ed29b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2650806
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Auto-Submit: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#848417}
GitOrigin-RevId: 11246e0d882eb72b21e04a4011f9404c0cfb4fd7