vaapi: Protect against potential protected session leak
There's a case that could occur where we get a Reset call during a
configuration change state which causes us to rebuild the accelerated
delegate. If the prior delegate has created a protected session, the new
one will not know about it and create a new one and leak the old one.

This also cleans up a case where swapping out the VaapiWrapper in a
delegate did not reset the HW key map, which should be done because it
is intended that a HW key map relates to a specific protected session.

BUG=b:178812146
TEST=Protected playback works

Change-Id: I5238fba0e68ada923c188fa954826c359cf9da8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2659738
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Cr-Commit-Position: refs/heads/master@{#850706}
GitOrigin-RevId: 0bab65aca7e009b02e54f50388bfdaf31723f4f6