media/gpu/v_e_a_tests: Lower tolerance of the quality for lower resolutions
VA-API encoder on Baytrail cannot produce a good quality stream
if the target bitrate is low, whereas a resolution of a video
to be encoded is low. This CL is a workaround to pass the video
test. The tolerance used in SSIM validator is set to 0.65 if a
resolution of a video is less than 360p, which is lower than
the default tolerance, 0.70.

Bug: b:170790027
Test: video_encode_accelerator_tests with 180p video on banjo
Change-Id: I7ede767268efb32efa41b6f62ae280ac96ce9218
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2679092
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#852006}
GitOrigin-RevId: f712dc9ca82508f01c32fe08fcd93aa3504e31ef