media: Compile AV1Decoder on Linux and LaCrOS.
This CL allows us to compile media::AV1Decoder in Linux and LaCrOS
builds. This will allow us to use the MSan and UBSan for fuzzers which
are only supported on Linux. Along these lines, we allow the VA-API
portion of the AV1 hardware decoding stack to compile on Linux since we
now have libva 2.10.0 in the sysroot.

Bug: b:175886979
Test: None
Change-Id: Iaf366289c2c1fd4e5097ddd0c648f39f4174bdc6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2680680
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: James Zern <jzern@google.com>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#853414}
GitOrigin-RevId: 6183e95ee75a81d1402a80d4d474b8d2eed13e14