Mac Capture: Restart capture if it hangs
Some cameras, when another process starts uses the camera, stop
producing frames in Chrome. No errors are reported, but all frames are
dropped (and no reason for the dropped frames is reported).

Add a callback that checks to see if a new frame has been produced. This
callback is made every 1 second. If 5 of these callbacks find that no
new frames have been produced, restart the capturer.

In local testing with FireFox and PhotoBooth as the other process, this
fixed Chrome and leaves the other process' capture working as well.

Bug: 1176568
Change-Id: I29df4bd95a591f815ecadaf862706950506bffa6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2686756
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Evan Shrubsole <eshr@google.com>
Reviewed-by: Henrik Boström <hbos@chromium.org>
Reviewed-by: Markus Handell <handellm@google.com>
Cr-Commit-Position: refs/heads/master@{#853811}
GitOrigin-RevId: 080b286230811e411997141b9ea33144bcd2b0ad