[Fuchsia] Fix OnFreeInputPacket handling in SysmemBufferWriterQueue
SysmemBufferWriterQueue::ResetPositionAndPause() is used to handle NoKey
error in FuchsiaSecureStreamDecryptor. It's expected to keep the queue
of input packets, so they can be re-sent when Unpause() is called. That
logic wasn't implemented correctly. Specifically it wasn't handling the
case when OnFreeInputPacket event was received before the stream is
restarted. This worked in the past because CDM wasn't releasing old
packets before the new stream is started, but that will be changed in
https://fxrev.dev/486837 . This change fixes the issue in
SysmemBufferWriterQueue to handle this case correctly. Also added
comments for some functions that are used in
FuchsiaSecureStreamDecryptor to handle the NoKey error.

Bug: 1179133
Change-Id: Ia385eb956ad09c193ef04876fb535fc273abd057
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2698152
Commit-Queue: Sergey Ulanov <sergeyu@chromium.org>
Commit-Queue: Yuchen Liu <yucliu@chromium.org>
Auto-Submit: Sergey Ulanov <sergeyu@chromium.org>
Reviewed-by: Yuchen Liu <yucliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#854947}
GitOrigin-RevId: df79dc4935862feda981ee68d2eff113a5c84240