Factor out utilities for media::VideoFrame => blink::Image.
This moves CreateImage from WebCodecs blink::VideoFrame into
platform/graphics where so it can be shared across core/ and
modules/ (we'll use it for <video> from ImageBitmap shortly)

This is mostly just a refactoring, though it did expose a bug
in GpuMemoryBuffer handling which is fixed. There were a fair
number of test fixture changes and refactoring necessary for
the tests to be written as well.

Bug: 1175907
Test: New tests. All existing tests pass.
Change-Id: I4857140d1d273f52aa6351655efb8d4aa941f1dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2708603
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Auto-Submit: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#856805}
GitOrigin-RevId: 19bb82998cd75b3c703441742987ba86cddfd65c