Adds AudioSourceFetcher to speech recognition service.
AudioSourceFetcher can be used to access microphone data for
transcription. It creates and owns a SpeechRecognitionRecognizer
to which it can pass audio, and is constructed with a StreamFactory
from which it can get an AudioSource.

The goal of AudioSourceFetcher is to abstract the work of listening
to mic audio from the client of the speech recognition service and
push getting audio data closer to the SODA service, making it easier
for clients to access transcriptions of mic audio and decreasing
number of layers/hops for audio to get to SODA. AudioSourceFetcher
on Chrome (and web speech fallback) will run in the utility process,
while on Chrome OS it runs in the browser process.

This change sets up the skeleton AudioSourceFetcher but does not
yet start listening to audio, which will happen in a future change.

See design at go/cros-dictation-with-soda.

Bug: 1173135
Change-Id: Ida84858bee01da31116f59d83c7b957f69e45a5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2715586
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Evan Liu <evliu@google.com>
Commit-Queue: Katie Dektar <katie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#859949}
GitOrigin-RevId: ca90aaa5b81a320d50d4dcf1b97495046400fe32