vaapi: Fix protected HEVC multi-slice corruption
When we have multiple segments, we were not specifying the IV for a
segment if it was clear. Apparently the Intel implementation needs that
information (which makes sense, it uses the one from the first segment
for all the segments and then tracks the CTR counter itself). Specifying
the IV for clear segments fixes the problem.

BUG=b:182210113
TEST=All clips listed for repro play w/out corruption in Netflix

Change-Id: I1031e9e74d52c97f59b1140e59e82b990a0a7d6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2743456
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Auto-Submit: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#860989}
GitOrigin-RevId: a749abd453c6ae421dd9022fbf259fb721949e87