Handle Web Audio's "playback" latency profile properly in AAudio back-end
The "playback" latency profile falls into
AAUDIO_PERFORMANCE_MODE_POWER_SAVING. This mode uses a huge
underlying buffer capacity (5766) and comes with relatively
irregular callback timing.

The buffer size mismatch caused the worklet rendering mode
to glitch because of the thread hop from the AudioOutputDevice
thread to the RealtimeAudioWorkletThread.

The solution is to use a bigger buffer size (6400) on the
Web Audio client (AudioDestination) to match the underlying
buffer size.

This will increase the latency of the "playback" profile to
~120ms.

Bug: 1181434
Test: The reported repro case does not cause glitches.
Change-Id: Ieddca00647650ed2d61717e0b9cc64605410ec0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2737361
Reviewed-by: Olga Sharonova <olka@chromium.org>
Reviewed-by: Thomas Guilbert <tguilbert@chromium.org>
Commit-Queue: Hongchan Choi <hongchan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#862046}
GitOrigin-RevId: f7043f2874e35407775e5daba0912cb2ffb76f14