AudioSourceFetcherImpl can listen to audio streams.
Adds and implements AudioSourceFetcher::Start, and the audio
capture callback which converts audio data into mojom and sends
it to the SpeechRecognitionRecognizer.

Refactors audio conversion work, which is common between
ChromeSpeechRecognitionClient and AudioSourceFetcher, into
AudioDataS16Converter.

For more, see go/cros-dictation-with-soda.

This completes end-to-end functionality for OnDeviceSpeechRecognizer
which can be used by Projector and Dictation to use on-device
speech recognition.

There are some follow-ups needed to pass audio parameters from
OnDeviceSpeechRecognizer to AudioSourceFetcher, left as TODOs.

Bug: 1173135
Change-Id: I460034d780ab07010b7a23c8dab018e6d4754a1f
AX-Relnotes: N/A
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2740120
Commit-Queue: Katie Dektar <katie@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Evan Liu <evliu@google.com>
Cr-Commit-Position: refs/heads/master@{#862219}
GitOrigin-RevId: 92043393b03f4002f11319eb9aff7b1bdfdafe76