[Fuchsia] Fix FuchsiaAudioRenderer to call Stop() only after Start()
AudioConsumer expects Stop() to be called after Start().
FuchsiaAudioRenderer uses |state_| field to track the current state.
Previously it was also used to track end-of-stream state, but
end-of-stream packet may be processed without the stream being started.
FuchsiaAudioRenderer::StartTicking() was assuming that state_!=kStopped
indicates that the stream is started, but that wasn't always the case.
Added a new boolean field to track end-of-stream, separately from
started/stopped state, which allows to ensure that Stop() is never sent
before Start().

Bug: 1177897
Change-Id: I4fb8926d4283329f2838a8cee658d8f38e396cb5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2769018
Commit-Queue: Sergey Ulanov <sergeyu@chromium.org>
Reviewed-by: David Dorwin <ddorwin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#864101}
GitOrigin-RevId: 67549d4ebe15799b424c62510ac2c7e0df312051