media/gpu/vaapi/test Add AV1 parser status enum
This change adds a ParserResult enum to the Av1Decoder class. This
is returned by the ReadNextFrame(), which performs parsing, not
decoding.

crrev.com/c/2740069 introduced an Av1Decoder for the VA-API decode test.
There was consideration to fold ReadNextFrame() into DecodeNextFrame();
this was decided against because doing so would not enable moving
obu_parser_ from an instance to a local variable.

However, ReadNextFrame() was returning a VideoDecoder::Result; this
doesn't make sense for ReadNextFrame() because it does not do decoding,
it just parses a frame, obtaining header data. This CL fixes this by
introducing an enum that is parsing-specific.

Testing details follow.
Using the videos in media/test/data and with a directory out_frames
existing,
      decode_test --video=test-25fps.av1.ivf --visible \
       --out-prefix=out_frames/25fps_frame
      decode_test --video=av1-film_grain.ivf --visible \
       --out-prefix=out_frames/film_frame

Bug: b:168064813
Test: run decode_test as described above
Change-Id: I1f571d12fffd11122fef080e46c6451610b2cdbc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2776036
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>
Commit-Queue: Clarissa Garvey <clarissagarvey@chromium.org>
Cr-Commit-Position: refs/heads/master@{#865846}
GitOrigin-RevId: c457c97a0f94094b4fe1d94c5bf78f7bd01628f2