media: Fix MojoCdmService creation crash
Currently MojoCdmService::Create() is a static function and the creation
is async. It also holds a raw pointer to the `context` which will be
used after creation is completed. However, during the creation process,
the owner of the `context` (e.g. CdmService) could have already been
destructed, causing a crash.

This CL fixes the issue by removing MojoCdmService::Create() and let
the owner of the `context` (e.g. CdmService) own the MojoCdmService
all the time, to avoid this issue.

Bug: 1190319
Test: Adds a new test that can repro this issue.
Change-Id: I97e25b396d4eb043fb9643c2053a86d13d6bb253
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2785849
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: John Rummell <jrummell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#866908}
GitOrigin-RevId: e9f8f5045f71da549250d516b7ff27bf865e3817