Make use of Optimized Scaling in HW encoder.
In this CL, we replace the ToI420()+libyuv::I420Scale() code path with
a CropAndScale()+GetMappedFrameBuffer() code path. This:
1. Avoids I420 conversion, e.g. if we are capturing in NV12.
2. This allows Optimized Scaling, i.e. mapping frames that are already
   of the correct size (in-capturer scaled frames).

Making use of CropAndScale() is also a prerequisite to zero-copy scaled
HW encoding (https://crbug.com/1194067) however in this CL we continue
to map the frame, which we already did prior to this CL in the ToI420()
call. The mapped frame gets copied by MojoVideoEncodeAccelerator in
order to pass it between processes.
- This is something that was already done prior to this CL, however in
  this CL, MojoVideoEncodeAccelerator is updated to support
  STORAGE_UNOWNED_MEMORY as is necessary with the new code path. We do
  this by reusing existing helper function
  MojoSharedBufferVideoFrame::CreateFromYUVFrame(), but we update it so
  that SHMEM regions are mapped prior to memcpy.

The new code path is only triggered when Optimized Scaling (behind flag)
is enabled, otherwise we fall back to the old path. But when Optimized
Scaling has rolled to 100% we can remove the old path (todo added).

Bug: chromium:1193457
Change-Id: I74ab5bcc8b86fa464df626e590f2b160585a62e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2794483
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Eugene Zemtsov <eugene@chromium.org>
Reviewed-by: Evan Shrubsole <eshr@google.com>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Auto-Submit: Henrik Boström <hbos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#868660}
GitOrigin-RevId: 5cc8e4293785349f9312a287ae1a9a43d9e7f7a2