media/gpu/v4l2: add VP9 stateless accelerator for 5.4 kernel
Add a VP9 accelerator based on the latest (but not final) upstream
proposal that has been merged in the 5.4 tree.

This accelerator is specific to Chromium-OS and should be replaced by
the final upstream one when it lands.

BUG=b:161415741
BUG=b:167469080
TEST=VDAtest passes with VP9 input on Asurada.

Change-Id: I408534f53d1f23bf7f22b0ba4a3b7532d437d08f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2434185
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Auto-Submit: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#869062}
GitOrigin-RevId: 2ed1f558139fa1f9f354b9b6c16ef3ec952cf5a1