Implement support for alpha preference in WebCodecs.
This change adds an alpha={keep,discard} preference to VideoFrameInit
and VideoEncoderConfig. On VideoFrameInit the default is 'keep' while
on VideoEncoderConfig the default is 'discard'. The defaults were
picked based on expected use cases.

In VideoFrame construction we use the alpha preference to force the
media::VideoPixelFormat from transparent to opaque. In the future we'll
use this to inform the type of snapshot we get from CanvasImageSource.

In VideoEncoderConfig we currently throw a 'NotSupportedError' if
alpha = 'keep' is specified. Some later change will add support for
alpha encoding

R=chcunningham, eugene

Fixed: 1192705
Change-Id: I711e33a61767b74566df871fe260db404ea05a77
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2802760
Reviewed-by: Eugene Zemtsov <eugene@chromium.org>
Reviewed-by: Chrome Cunningham <chcunningham@chromium.org>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#869292}
GitOrigin-RevId: 322a176b4ef25e644808dd1748b6a63a2764e21d