Add mojo API for changing the live caption language
This CL adds an API to the interface between the renderer and the speech recognition service that will be used to change the language of the Speech On-Device API (SODA) mid-stream. The config path will be verified and passed to the renderer from the browser process.

Bug: 1175357
Change-Id: I3bfa4a43ea5fbac89fb2818892e352188f711d0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2691532
Commit-Queue: Evan Liu <evliu@google.com>
Reviewed-by: Will Harris <wfh@chromium.org>
Reviewed-by: Rob Schonberger <robsc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#870219}
GitOrigin-RevId: 6dc7117ed3cd4e8e236cfb71cd4b89d94d15e759