media/vaapi: don't call vaSyncSurface() during encoding on Intel
vaSyncSurface() after vaBegin-Render-EndPicture takes a significant
amount of time as per analysis in the bug; this call seems unnecessary
as long as there is a vaMapBuffer() following, and it might introduce
some unintended side effects when several contexts are active, so
let's remove it for Intel, and leave it for AMD, where I've seen
encoding breaking without it. I verified it on samus (BDW, H264) and
dedede (JSL, VP9) with appr.tc loopbacks.

Bug: b:184312032
Change-Id: I158981c189373524af5309df11f9c2dcf200470e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2803050
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Sreerenj Balachandran <sreerenj.balachandran@intel.com>
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#871742}
GitOrigin-RevId: b4c088292c13e97f62c7d6175f1b701a789b4bcb