Fix waiting for decoder status in SoftwareVideoEncoderTest
RunUntilIdle() was used to wait for the status of a
VideoDecoder::Decode() call. For certain decoder implementations this
can return too early because we don't wait for tasks that execute in
system queues.

The fix is to wait for the event we're interested in explicitly, as
recommended by the documentation of TaskEnvironment::RunUntilIdle().

Change-Id: I56eb72f119fa9cc3cdd2e4cd2e6903d565bd157a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2821539
Reviewed-by: Eugene Zemtsov <eugene@chromium.org>
Commit-Queue: Wojciech Dzierżanowski <wdzierzanowski@opera.com>
Cr-Commit-Position: refs/heads/master@{#871885}
GitOrigin-RevId: 007a1d378ec9e24373285f78d1f4cd7b79c83bd3