media/gpu/v4l2: do not fail if buffer cannot be cropped exactly as requested
Some hardware may not be able to accomodate all crop requests: for
instance, it is not uncommon for video hardware to only be able to work
with even dimensions. This is not a problem for us as long as the
cropped rectangle is larger than the requested one: since we will pass
the actual visible rectangle to the rest of the pipeline, the display
hardware will perform the proper cropping itself.

BUG=b:184913721
BUG=b:185734284
BUG=b:185733582
TEST=video.DecodeAccel.vp8_odd_dimensions passes on Asurada.

Change-Id: I9ca47dee1de73c8173588cf736ee4179ae77d0c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2843035
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Auto-Submit: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#874984}
GitOrigin-RevId: e8e0c486db021f6477eed8dfaec7c120292d4a7c