SharedImage: Use enum for plane identifier
In the previous patch, the plane was passed as a uint32_t. This started
becoming confusing while writing subsequent patches. Use an enum,
gfx::BufferPlane, instead.

Also also put the gfx::BufferPlane argument after the
gfx::BufferFormat argument in the various functions that take both, so
that it is more pretty.

Also clean up the error checking. Do checks for
  plane != gfx::BufferPlane::DEFAULT
anywhere that we don't support non-default plane binding, which is in
the individual factories. Check for
  IsPlaneValidForGpuMemoryBufferFormat
to validate that arguments are sensical (e.g, we are not requesting
the Y plane of a RGBA texture).

Bug: 1201865
Change-Id: I2b943b737e3d498867cde576bf75b1e644e789f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2849354
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: ccameron <ccameron@chromium.org>
Commit-Queue: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#876292}
GitOrigin-RevId: 620140af3d434661b16550f2c188621dc84e41f8