media/gpu/v4l2vd: move call to CompleteFlush()
When a flush fence is met, the output request queue is expected to be
empty as suggested by a DCHECK(). It is thus safe to move the call to
CompleteFlush() into the same code block instead of running it just a
bit later as we will break the loop. This improves readability (all the
flush-related code is in the same place) and will allow us to use
resume_decode in non-flush situations.

BUG=b:185734149
TEST=tast video.DecodeAccelVD.vp9_show_existing_frame passes on Asurada.

Change-Id: Ie07257d72e76f14879013d22e7f1a9edddeef8c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2856123
Auto-Submit: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Fritz Koenig <frkoenig@chromium.org>
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#877056}
NOKEYCHECK=True
GitOrigin-RevId: f577a9c1773f56c2696336d87989767ae8cb1593