media/gpu/v4l2vd: schedule decode work if next surface to output is not decoded yet
It may happen that the next surface to output is already in the output
queue while its decoding work has not been submitted yet. When this
happens, avd_->Decode() must be called to submit the decoding job to the
kernel, or we risk to make the decoder thread starve.

Make sure this happens by scheduling a call to DoDecodeWork() that will
continue the decoding job.

BUG=b:185734149
TEST=tast video.DecodeAccelVD.vp9_show_existing_frame passes on Asurada.

Change-Id: I1d62810cd8fd2e2ca7a4c93796284465fa3cc05a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2856124
Auto-Submit: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Cr-Commit-Position: refs/heads/master@{#877394}
NOKEYCHECK=True
GitOrigin-RevId: 3114ac603a53b0031279d07e4f9ed423be76795f