Move VideoFrameYUVConverter::VideoFrameYUVMailboxesHolder to own class
This utility class has functionality that is very useful for accessing
the planes of VideoFrames as SkImages and SkSurfaces. Separate it out,
so this functionality can be put in a single useful place.

This class also has a separate functionality to create a copy of a
VideoFrame on a given RasterProvider (by copying through the CPU).
This probably belongs in a separate class, and will be moved out
next.

Bug: 1206168
Change-Id: I263b642beb81cc00cc9ab3e7831e100d70769580
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2881030
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#881440}
NOKEYCHECK=True
GitOrigin-RevId: fbf39d1c6c8f6243edfc88cb7165b378de645e45