media: Make IMFCdmProxy a C++ interface
Currently IMFCdmProxy is a Windows COM interface. But it's really just
a normal class used internally in Chromium, and it doesn't need to be a
COM interface. Actually it's the only COM interface we define in
Chromium (other than in third_party or test).

This CL changes IMFCdmProxy to be a normal C++ interface. We also rename
it to MediaFoundationCdmProxy to match the interface type.

Bug: 1196826
Change-Id: I18609be6c5e8e6c623217a2d6c64d78c3176eff4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2889741
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: John Rummell <jrummell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#883079}
NOKEYCHECK=True
GitOrigin-RevId: ce34ac4a0561e1f7e24dccec8e22f2c6fa0f3b83