Add feature flag to allow WGC for desktop capture.
This change introduces a feature flag that controls whether we can use
the new DesktopCapturer implementation based on the
Windows.Graphics.Capture APIs. This implementation is only available on
Windows 10 RS5 and up, so even if it is allowed, it may not be used all
the time.

When it is used, it will add a yellow border around the window or
screen that is being captured. This is why we only want to enable this
feature in desktop_capture_device.cc and not
native_desktop_media_list.cc, as using this capturer for thumbnails
will outline every window in yellow and lead to an overwhelming
experience.

I also updated the capturer creation code to set the
allow_cropping_capturer option to true and call
DesktopCapturer::CreateWindowCapturer instead of directly invoking
CroppingWindowCapturer. This will not change the functionality
unless a higher priority option is set, and allow_wgc_capturer is
the only one.

Bug: webrtc:9273
Change-Id: I6035462611641b40dcb82ee222ff2b1196fa0b9a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2855288
Reviewed-by: Joe Downing <joedow@chromium.org>
Reviewed-by: Olga Sharonova <olka@chromium.org>
Commit-Queue: Austin Orion <auorion@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#884633}
NOKEYCHECK=True
GitOrigin-RevId: 4d1b9211722f6e38feac9476498692caa75a0af9