Migrate media/blink/ off of ThreadTaskRunnerHandle::Get()
This is a precursor CL before moving media/blink/ to Blink, where the
usage of ThreadTaskRunnerHandle::Get() is banned.

This CL changes most classes to take a task runner parameter which is
properly selected at the call sites. In the case of unittests, the task
runner is provided by BlinkPlatformWithTaskEnvironment, which is the
class in charge of setting up Blink and the TaskEnvironment for the
tests.

Bug: 1198361
Change-Id: Ic98eaa01b31ee46289679869a6ff26ed1df93b09
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2896318
Commit-Queue: Henrique Ferreiro <hferreiro@igalia.com>
Reviewed-by: Will Cassella <cassew@google.com>
Cr-Commit-Position: refs/heads/master@{#886483}
NOKEYCHECK=True
GitOrigin-RevId: 3767cbf84ab0a1ed9b17b80cf31094f697c30ca7