Ensure JPEG decode tasks are executed on decoder_thread_
Some of the JPEG decode tasks are not running on the decoder_thread_ so,
the sequence checkers in vaapi_wrapper (CL:2793005) are crashing
during JPEG decode scenarios. This CL ensures necessary JPEG decode
tasks are executed on the decoder_thread_. In addition, we are
refactoring MJPEG stack to use asynchronous initialization.

BUG=b:183515581
TEST=JPEG decode shouldn't cause any crashes

Signed-off-by: Sushma Venkatesh Reddy <sushma.venkatesh.reddy@intel.com>
Change-Id: I5854f8ae5c0a8a213b3a178d1a4facb00422e908
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2808778
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#887459}
NOKEYCHECK=True
GitOrigin-RevId: 900b973eaa7bebcac33f1df56b488a8db7abc98e