Add support to WebM parser for the "projection" element
This change modifies several classes related to the WebM parser in
order to support the projection element defined in the Matroska
spec. Without this, any video stream containing this element will
fail to parse causing the video to fail to play.

The projection element contains an element called "Projection Private".
This element is quite complex. This change does not add the logic
to parse the data from this element since the main goal is to
enable playback of these videos. Instead I simply tagged it with
SKIP_BINARY.

https://www.matroska.org/technical/elements.html#Projection

Fixed: 1210299
Change-Id: Iae0cd57be670362fdc80d8f632c8747824ac105c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2903858
Commit-Queue: Matthew Wolenetz <wolenetz@chromium.org>
Reviewed-by: Matthew Wolenetz <wolenetz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#888101}
NOKEYCHECK=True
GitOrigin-RevId: 86a73f2bf8d3b2bba7d481710639e0a070b339c6