media/gpu/test: Deprecate video thumbnail tests.
The thumbnail test was introduced as a workaround for devices that
don't support import mode, which is a requirement for the frame
validator to work. Since the last devices that don't support import
mode had their final release in R91, we can now remove all code related
to the thumbnail test.

Bug: b:172225784
Test: ./video_decode_accelerator_tests on hatch
Change-Id: I551ac2f7eb401c2ca92aac44104ae74a661ad07f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2750844
Commit-Queue: David Staessens <dstaessens@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#888715}
NOKEYCHECK=True
GitOrigin-RevId: 7c5ad27d2758e284d0ceedab7ee67a8178b79b01