Move media/blink/ to Blink
media/blink/ files are being moved to public/platform/media/ and
renderer/platform/media/, depending on the private/public status of the
file as mentioned in this table:
https://docs.google.com/document/d/1p0OuzArDNZHT_aOm75oOWh3qR-mT6Zq-qKxxKYT-jYI/edit#bookmark=kix.2v0uvunkva0g

The move includes renaming files as described in the first table at
https://docs.google.com/document/d/1p0OuzArDNZHT_aOm75oOWh3qR-mT6Zq-qKxxKYT-jYI/edit#heading=h.jqrzzxjh40df
where underscores are added to be consistent with the camel-case naming
of the classes defined in the renamed files.

//media/blink:media_blink_unittests is split with all files but one
being moved to blink_platform_unittests.
web_media_player_impl_unittest.cc uses WebView, so it requires a
Platform object and, as such, should be moved into
//third_party/blink/renderer/platform:unit_tests which is part of
blink_unittests. This makes the android bots fail, as they find no tests
to run. Because of this, this CL leaves this test in place, which will
be removed in a follow-up, together with all of media/blink/.

Only the header guards and the export macros are updated. Namespace
changes and other Blink coding conventions (GURL => KURL, Blink mojom
includes, etc.) are left for follow-ups. Because of the latter, this CL
fails the Blink-variant mojom presubmit check.

No-Presubmit: true
Bug: 1198341
Change-Id: I66e46c970db1baae453d16da738fa3507578a4f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2847359
Reviewed-by: Will Cassella <cassew@google.com>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Henrique Ferreiro <hferreiro@igalia.com>
Cr-Commit-Position: refs/heads/master@{#890853}
NOKEYCHECK=True
GitOrigin-RevId: b56c921dc8c64ae3d7b5a3a04b168e7773dc77d6