gfx::ColorTransform: Add options argument
The existing intent argument had the same behavior for all intents
except the intent that isn't an intent, which is the intent to disable
optimizations.

Change this to be an optional options structure. Move the bit depth
arguments into this structure. The options will soon get tonemapping
and sdr white level parameters.

Bug: 1204888
Change-Id: I97c9b20a871051365e072675a55c463c2eaebae6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2910939
Reviewed-by: ccameron <ccameron@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#891430}
NOKEYCHECK=True
GitOrigin-RevId: 3c71b66a5222492abbae8f5f1eb5980c131887d1