GpuMemoryBufferVideoFramePool: Parameterize usage
Current use of GpuMemoryBufferVideoFramePool is to create GMB-backed
frames that are populated by writing using the CPU, and so the
gfx::BufferUsage is hard-coded to gfx::BufferUsage::
SCANOUT_CPU_READ_WRITE. For one copy canvas capture, out CPU access
to the GMBs will be read-only, so we'll want to use the
usage SCANOUT_VEA_CPU_READ, which is what is currently used by
video capture frames.

Also update macOS-GMB support to allow SCANOUT_VEA_CPU_READ, since
that configuration is valid (and is already in use, just not in a way
that hits this path).

Bug: 1207111
Change-Id: I82e1a0d16c92748db8a8dc1ed118bcebec723be4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2942208
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Reviewed-by: ccameron <ccameron@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#891790}
NOKEYCHECK=True
GitOrigin-RevId: 742c88c8ea62f3a7b63a9a6feba2d317b72ee938