media: Add support for AMD transcrypted content
This adds protected content support for AMD. AMD does this by
transcrypting the content first via OEMCrypto which normalizes it to a
fully encrypted block. We do that part by using the Decryptor from the
CDM just like decrypt to clear. We are returned the transcrypted content
plus the key material needed to decrypt that by the video decoder. In
addition, there is a clear lead returned which is all the headers
required by the codec parser. (This only implements CENCv3 for now).

This is handled by doing the transcryption in a new class
DecoderBufferTranscryptor which is owned by the ChromeOS
VideoDecoderPipeline first and then passing the result of that
into the decoder for normal processing. Then in the accelerators we just
need to reinject the key material so the HW decoder can decrypt before
decode.

Corresponding ChromeOS CL: https://crrev.com/i/3833522

BUG=b:177691210
TEST=Volteer protected content works, majolica protected content works
     with full prototype.

Change-Id: I8f0f9a6e7c33fcb31040546f31eb3811615dbaac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2900427
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#893595}
NOKEYCHECK=True
GitOrigin-RevId: efff52a0eee4b2a131cde77e51331f4dc535b059