media: Convert to DCHECK for curr_decode_task on resolution change
From a prior code review, it came up that |current_decode_task_|
should always exist because we are ensuring the |state_| is
kChangingResolution, which only happens with an active task. If we got
reset, the state would be changed. Convert it from a conditional check
to a DCHECK to signify this.

BUG=None
TEST=None

Change-Id: Icdd55374866453bb8a2bd022d348e5db88edfcf2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2983661
Auto-Submit: Jeffrey Kardatzke <jkardatzke@google.com>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#895408}
NOKEYCHECK=True
GitOrigin-RevId: 634bf05afbaa2e911271147355a1fe76a349498f