[Fuchsia] Improve underflow handling in FuchsiaAudioOutputDevice
Previously FuchsiaAudioOutputDevice was checking for underflows when
scheduling PumpSamples() timer and then in PumpSamples() itself. When
trying to recover it was also delaying the next packet by
kLeadTimeExtra, which is not necessary. This change moves underflow
logic to PumpSamples(), which allows to simplify that logic and reduce
the effects of missed timer deadline.

Bug: 1066203
Change-Id: I617ec7f7a248360fd0a17575d7f36f59f7f77eb0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2989095
Auto-Submit: Sergey Ulanov <sergeyu@chromium.org>
Reviewed-by: Kevin Marshall <kmarshall@chromium.org>
Commit-Queue: Kevin Marshall <kmarshall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#896121}
NOKEYCHECK=True
GitOrigin-RevId: d401b74a149e1108952f4e79263e76beeddf4c07