Add decoder type query to gpu_mojo_media_client
In cases where MojoVideoDecoder was able to pre-determine that the
remote decoder could not be used due to mismatched profiles, we'd like
to still report which decoder was unable to be initialized. In this
case, we have to ask |GpuVideoAcceleratorFactories|, which can go off to
ask GpuMojoMediaClient.

Change-Id: If9065e57b01f781c024cd6b111361d2bc1a817f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3003692
Reviewed-by: Frank Liberato <liberato@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Ted Meyer <tmathmeyer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#899871}
NOKEYCHECK=True
GitOrigin-RevId: dfb2729b75acad1d725ed8d3c2495ffcf8e26c82