media/gpu/test/video_encoder_client: Remove wrong DCHECK
We check the encoder client state is not kIdle in EncodeDoneTask().
The code assumes encoding stops in the last frame to be encoded,
that is, EncodeDoneTask() for the last frame is called. It is now
wrong because encoding stops in the middle in ForceKeyFrame,
DynamicBitrate/FramerateChange test cases. This removes the wrong
DCHECK.

Bug: None
Test: video_encode_accelerator_tests with dcheck_always_on=true
Change-Id: I038f14ea7c1c33b0c636ab8f14e8c993abf82869
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3060860
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: David Staessens <dstaessens@chromium.org>
Auto-Submit: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: David Staessens <dstaessens@chromium.org>
Cr-Commit-Position: refs/heads/master@{#906986}
NOKEYCHECK=True
GitOrigin-RevId: 1e7874bd5f81ed9a277f9d49ec99d472b80f847f