media/gpu/chromeos: make sure VideoDecoderMixins respect Decode()
media::VideoDecoder::Decode() say that "|decode_cb| will not be called
from within this method" [1]. This CL makes that the case in the two
impls (Vaapi|V4L2)VideoDecoder.

This was raised in crrev.com/c/3049659. Note that there was another
concern there re. |decode_cb| ""... and that it will be called even
if Decode() is never called again.". These seems to be fulfilled
in both implementations (see e.g. [2] for Vaapi). (I would also argue
that if this was to be enforced, the interface should use
ScopedClosureRunner.)

[1] https://source.chromium.org/chromium/chromium/src/+/main:media/base/video_decoder.h;l=85-87;drc=5db067a3dc38dac442279730580c608d1db5e709
[2] https://source.chromium.org/chromium/chromium/src/+/main:media/gpu/vaapi/vaapi_video_decoder.cc;l=131?q=VaapiVideoDecoder::~VaapiVideoDecoder

Bug: b:192087509
Change-Id: Ie6d697d5e1997416344e2cea58c5cbc23584cc94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3055391
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#907291}
NOKEYCHECK=True
GitOrigin-RevId: e4bb5ae4828bedf63b9f5404abd88e6235eb5a8b