media: Limit # of concurrent VaapiVideoDecoders.
This CL limits the number of VaapiVideoDecoder instances that can exist
at once. This is to avoid running out of FDs in the GPU process.

The soft limit of FDs is 2048. I experimented playing 32 videos at once
for different codecs to see how many FDs would be open when I impose a
limit of 16 VaapiVideoDecoder instances (that means that 16 of the 32
videos use software decoding). Here are the results on a volteer when
all 32 videos are playing at once:

- With an H.264 video with a max. of 16 reference frames:

  The number of FDs is ~1699.

- With a VP8 video:

  The number of FDs is ~925.

- With a VP9 video:

  The number of FDs is ~1180.

- With an AV1 video that uses film grain synthesis (which needs more
  buffers):

  The number of FDs is ~1970.

Bug: b:181264362
Test: Play 32 videos at once and monitor # of GPU process FDs.
Change-Id: Ibf55682ef38b3639382150f76805f2cac33b1695
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3046947
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909165}
NOKEYCHECK=True
GitOrigin-RevId: 6d7b52b2987587a7de85939b87f4ccbbc2d466fd