media: Refactor Direct Composition Surface Handle Registration
Instead of sending the direct composition surface handle from the
MediaFoundationService (utility/LPAC) process to the render process,
then register it in the GPU process via the browser process, we send the
handle from the MediaFoundationService process to the browser process,
which will forward the handle to the GPU process to be registered.

This eliminates the security risk of sending an arbitrary handle from
the low-privileged render process to the GPU process.

Also updated mojom interfaces to return an optional UnguessableToken,
because it's illegal to pass empty tokens across processes.

Bug: 999747
Test: Manually tested playback.
Change-Id: Ia8811010208c9a6ef7c5423754c0672ef050a6cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3070658
Auto-Submit: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Reviewed-by: Frank Li <frankli@microsoft.com>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#909317}
NOKEYCHECK=True
GitOrigin-RevId: 978d52421f6bc9fa61a768f2c9de75e3940e53e2